+++ podchat/src/main/java/com/fanap/podchat/call/model/ClientDTO.java
@@ -12,6 +12,10 @@ public class ClientDTO implements Parcelable {
     private String brokerAddress;
     private String desc;
     private String sendKey;
+    private String sendMetaDataTopic;
+    private String reciveMetaDataTopic;
+    private Boolean mute;
+    private Boolean video;
 
     public String getClientId() {
         return clientId;
@@ -41,9 +45,9 @@ public class ClientDTO implements Parcelable {
         if (topicSend != null && topicSend.endsWith(VIDEO_TOPIC_SUFFIX))
             return topicSend;
         if (topicSend != null && topicSend.endsWith(VOICE_TOPIC_SUFFIX))
-            return topicSend.replace(VOICE_TOPIC_SUFFIX,VIDEO_TOPIC_SUFFIX);
+            return topicSend.replace(VOICE_TOPIC_SUFFIX, VIDEO_TOPIC_SUFFIX);
         if (topicSend != null)
-            return VIDEO_TOPIC_SUFFIX +  topicSend;
+            return VIDEO_TOPIC_SUFFIX + topicSend;
         return null;
     }
 
@@ -51,9 +55,9 @@ public class ClientDTO implements Parcelable {
         if (topicSend != null && topicSend.endsWith(VOICE_TOPIC_SUFFIX))
             return topicSend;
         if (topicSend != null && topicSend.endsWith(VIDEO_TOPIC_SUFFIX))
-            return topicSend.replace(VIDEO_TOPIC_SUFFIX,VOICE_TOPIC_SUFFIX);
+            return topicSend.replace(VIDEO_TOPIC_SUFFIX, VOICE_TOPIC_SUFFIX);
         if (topicSend != null)
-            return  VOICE_TOPIC_SUFFIX + topicSend;
+            return VOICE_TOPIC_SUFFIX + topicSend;
         return null;
     }
 
@@ -63,7 +67,7 @@ public class ClientDTO implements Parcelable {
         if (topicReceive != null && topicReceive.endsWith(VOICE_TOPIC_SUFFIX))
             return topicReceive.replace(VOICE_TOPIC_SUFFIX, VIDEO_TOPIC_SUFFIX);
         if (topicReceive != null)
-            return  VIDEO_TOPIC_SUFFIX + topicReceive;
+            return VIDEO_TOPIC_SUFFIX + topicReceive;
         return null;
     }
 
@@ -73,13 +77,13 @@ public class ClientDTO implements Parcelable {
         if (topicReceive != null && topicReceive.endsWith(VIDEO_TOPIC_SUFFIX))
             return topicReceive.replace(VIDEO_TOPIC_SUFFIX, VOICE_TOPIC_SUFFIX);
         if (topicReceive != null)
-            return  VOICE_TOPIC_SUFFIX + topicReceive;
+            return VOICE_TOPIC_SUFFIX + topicReceive;
         return null;
     }
 
-    public String getBrokerAddress() {
-        return brokerAddress;
-    }
+//    public String getBrokerAddress() {
+//        return brokerAddress;
+//    }
 
     public void setBrokerAddress(String brokerAddress) {
         this.brokerAddress = brokerAddress;
@@ -101,6 +105,37 @@ public class ClientDTO implements Parcelable {
         this.sendKey = sendKey;
     }
 
+    public String getSendMetaDataTopic() {
+        return sendMetaDataTopic;
+    }
+
+    public void setSendMetaDataTopic(String sendMetaDataTopic) {
+        this.sendMetaDataTopic = sendMetaDataTopic;
+    }
+
+    public String getReciveMetaDataTopic() {
+        return reciveMetaDataTopic;
+    }
+
+    public void setReciveMetaDataTopic(String reciveMetaDataTopic) {
+        this.reciveMetaDataTopic = reciveMetaDataTopic;
+    }
+
+    public Boolean getMute() {
+        return mute;
+    }
+
+    public void setMute(Boolean mute) {
+        this.mute = mute;
+    }
+
+    public Boolean getVideo() {
+        return video;
+    }
+
+    public void setVideo(Boolean video) {
+        this.video = video;
+    }
 
     @Override
     public int describeContents() {
@@ -115,6 +150,13 @@ public class ClientDTO implements Parcelable {
         dest.writeString(this.brokerAddress);
         dest.writeString(this.desc);
         dest.writeString(this.sendKey);
+
+        dest.writeString(this.sendMetaDataTopic);
+        dest.writeString(this.reciveMetaDataTopic);
+//        dest.writeBoolean(this.mute);
+//        dest.writeBoolean(this.video);
+
+
     }
 
     public ClientDTO() {
@@ -127,6 +169,12 @@ public class ClientDTO implements Parcelable {
         this.brokerAddress = in.readString();
         this.desc = in.readString();
         this.sendKey = in.readString();
+        this.sendMetaDataTopic = in.readString();
+        this.reciveMetaDataTopic = in.readString();
+//        this.video = in.readBoolean();
+//        this.mute = in.readBoolean();
+
+
     }
 
     public static final Parcelable.Creator<ClientDTO> CREATOR = new Parcelable.Creator<ClientDTO>() {